More documentation on internal traits
